let PDMS/E3D load our UIC files through the CAF_UIC_PATH mechanism
This adds an additional step to the installation procedure but allow users to influence the tab order in E3D (which isn't possible when we add the UIC files during addin startup). Also: not all applications in E3D have ribbons, so we need to load different UIC files based on started application anyways.